Renaming file to match class name
