+++ lib/pact/consumer/mock_service/missing_interactions_get.rb
@@ -1,4 +1,4 @@
-require 'pact/consumer/mock_service/web_request_others'
+require 'pact/consumer/mock_service/web_request_administration'
 
 module Pact
   module Consumer
