+++ lib/pact/consumer/mock_service/pact_post.rb
@@ -1,5 +1,5 @@
 require 'pact/consumer_contract/consumer_contract_writer'
-require 'pact/consumer/mock_service/web_request_others'
+require 'pact/consumer/mock_service/web_request_administration'
 
 module Pact
   module Consumer
